代码与gitee同步，版本号V1.1.0

+++ src/monitor.lua
@@ -1,4 +1,9 @@
-#!/usr/bin/lua
+#!/usr/bin/env lua
+
+-- autor:jk110333@126.com 阿勇
+-- website:www.opoenwrtdl.coom
+-- 2022-07-06
+--
 socket = require"socket"
 cjson = require "cjson"
 lfs = require("lfs")
@@ -10,7 +15,16 @@ G_custom_conf={}
 function sleep(n)
     socket.select(nil, nil, n)
 end
+function get_shell_cmd_result(cmd)
+    local fp = io.popen(cmd)
+    local ret = ''
 
+    if fp then     
+        ret = fp:read("*all")
+        fp:close()
+    end
+    return ret
+end
 function timer_task_add(name,func_time)
     if name and type(func_time)=="table" and func_time.func and type(func_time.func) == "function" and func_time.period then
         func_time["task_info"] = {}
@@ -82,6 +96,7 @@ function cjson_decode_safe(str)
 end
 --read file all content
 function readfile(filename)
+	if not filename or #filename < 1 then return nil end
     local file = io.open(filename, "r")
     local data = nil
     if file then
@@ -164,9 +179,21 @@ function stop_conf_by_name(name)
     timer_task_stop(name)
     if name then
         G_custom_conf[name].stop = true
+		if G_custom_conf[name].type == "process" then
+			local pid = readfile(G_custom_conf[name].pidfile)
+			if pid and tonumber(pid) > 0 then
+				os.execute("kill -9 "..tostring(pid))
+			end
+		end
     else
         for n,conf in pairs(G_custom_conf) do
             conf.stop = true
+			if G_custom_conf[name].type == "process" then
+				local pid = readfile(G_custom_conf[name].pidfile)
+				if pid and tonumber(pid) > 0 then
+					os.execute("kill -9 "..tostring(pid))
+				end
+			end
         end
     end
     
@@ -182,7 +209,7 @@ function get_tab_len(t)
     return cnt
 end
 
-MY_VERSION="1.0.0"
+MY_VERSION="1.1.0"
 CTRL_RSP_FILE="/tmp/.lua-monitor.rsp"
 function print_conf_status()
     --writeData2file(CTRL_RSP_FILE,"lua monitor status,version "..MY_VERSION.."\n")
@@ -193,14 +220,18 @@ function print_conf_status()
     for name,conf in pairs(G_custom_conf) do
         local c = {}
         c.name = name
-        c.trigger_times = conf.trigger_times or 0
         c.type = conf.type
         c.stop = conf.stop or false
-        if conf.start_time then
+        c.trigger_times = conf.trigger_times or 0
+        if conf.start_time and c.stop == false then
             c.run_time =  os.time() - conf.start_time
         end
         c.last_state = conf.last_state 
-        appendData2file(CTRL_RSP_FILE,cjson.encode(c)) 
+		local pid = readfile(conf.pidfile)
+		if pid and tonumber(pid) > 0 then
+			c.pid = tonumber(pid)
+		end
+        appendData2file(CTRL_RSP_FILE,"    "..cjson.encode(c)) 
         if i < len then
             appendData2file(CTRL_RSP_FILE,",\n")
         end
@@ -216,7 +247,9 @@ function process_ctrl_msg(ctrl)
         for n, conf in pairs(G_custom_conf) do
             if conf.type == "process" then
                 local pid = readfile(conf.pidfile)
-                os.execute("kill -9 "..pid)
+				if pid and tonumber(pid) > 0 then
+                	os.execute("kill -9 "..tostring(pid))
+				end
                 debug("warn","kill process "..conf.name)
             end
         end
@@ -224,10 +257,13 @@ function process_ctrl_msg(ctrl)
         G_timer_tasks={}
         read_default_conf()
         walk_all_conf_json(G_conf.include)
+        writeData2file(CTRL_RSP_FILE,"reload done,all process task have restart")
+        timer_task_add("timer_logrotate",{func=task_cb_logrotate,period=600,args={}})
     elseif ctrl.cmd == "start" then
         if not ctrl.para or not G_custom_conf[ctrl.para] then
             debug("err","start conf name is err!")
-            os.exit()
+            writeData2file(CTRL_RSP_FILE,"start "..ctrl.para.." err")
+            return
         end
         start_conf_by_name(ctrl.para)
         writeData2file(CTRL_RSP_FILE,"start "..ctrl.para.." done")
@@ -235,7 +271,8 @@ function process_ctrl_msg(ctrl)
     elseif ctrl.cmd == "stop" then
         if not ctrl.para or not G_custom_conf[ctrl.para] then
             debug("err","start conf name is err!")
-            os.exit()
+            writeData2file(CTRL_RSP_FILE,"stop "..ctrl.para.." err")
+            return
         end
         stop_conf_by_name(ctrl.para)
         writeData2file(CTRL_RSP_FILE,"stop "..ctrl.para.." done")
@@ -255,12 +292,13 @@ function process_ctrl_msg(ctrl)
     
 end
 
+CTRL_MSG_FILE="/tmp/.lua-monitor.ctrl"
 --ctrl msg:
 --{"cmd":"reload/start/stop/stop_all/restart/start_all/status","para":"test"}
 function check_ctrl_msg()
-    CTRL_MSG_FILE="/tmp/.lua-monitor.ctrl"
     local json = readfile(CTRL_MSG_FILE)
     if json then
+        debug("info","read ctrl msg:"..json)
         local ctrl = cjson_decode_safe(json)
         if(type(ctrl) == "table") then
             process_ctrl_msg(ctrl)
@@ -286,9 +324,31 @@ function debug(level,s, ...)
     end
 end
 
+function generate_logrotate_conf(conf)
+    local logrotate_conf=""
+    os.remove("/etc/logrotate.d/"..conf.name)
+    if not conf["log_file"] or conf["log_file"] == "/dev/null" then
+        return
+    end
+    
+    logrotate_conf=conf.log_file.."\n{\n\tdaily\n\tmissingok\n\tcopytruncate"
+    logrotate_conf=logrotate_conf.."\n\tsize="..(conf.log_size or "300K")
+    logrotate_conf=logrotate_conf.."\n\trotate "..tostring(conf.log_cnt or 5)
+    if conf.log_compress ~= false then
+        logrotate_conf=logrotate_conf.."\n\tcompress"
+    end
+    logrotate_conf=logrotate_conf.."\n\tsu root root"
+    logrotate_conf=logrotate_conf.."\n}\n"
+    writeData2file("/etc/logrotate.d/"..conf.name,logrotate_conf)
+end
+
 function task_cb_process(conf)
     if not G_custom_conf[conf.name]["pid"] then
         G_custom_conf[conf.name]["pid"] = tonumber(readfile(conf.pidfile))
+    end
+    if not G_custom_conf[conf.name]["has_generate_logrotate_conf"] then
+        generate_logrotate_conf(conf)
+        G_custom_conf[conf.name]["has_generate_logrotate_conf"] = true
     end
     --print("check process stat file ".."/proc/"..tostring(G_custom_conf[conf.name]["pid"]).."/stat")
     local stat = readfile("/proc/"..tostring(G_custom_conf[conf.name]["pid"]).."/stat")
@@ -305,12 +365,20 @@ function task_cb_process(conf)
         end
     end
     --kill old
-    os.execute("kill -9 "..tostring(G_custom_conf[conf.name]["pid"]))
+	if G_custom_conf[conf.name]["pid"] then
+    	os.execute("kill -9 "..tostring(G_custom_conf[conf.name]["pid"]))
+	end
     --process has die,restart it in background
-    os.execute(G_custom_conf[conf.name]["cmd"].." &")
+    if G_custom_conf[conf.name]["log_file"] then
+        os.execute(G_custom_conf[conf.name]["cmd"].." >> "..G_custom_conf[conf.name]["log_file"].." 2>&1 &")
+    else
+        os.execute(G_custom_conf[conf.name]["cmd"].." > /dev/null 2>&1 &")
+    end
+    
     G_custom_conf[conf.name]["trigger_times"] = (G_custom_conf[conf.name]["trigger_times"] or 0) + 1
     G_custom_conf[conf.name]["start_time"] = os.time()   --启动时间
     G_custom_conf[conf.name]["pid"] = nil
+
     debug("warn","process "..conf.name.." restart!")
 end
 
@@ -350,18 +418,27 @@ function task_cb_file(conf_tab)
 end
 function task_cb_dir(conf_tab)
     local state = ""
-    if conf_tab.check == "modification" then
-        local file_attr = lfs.attributes(conf_tab.path)
-        if type(file_attr) == "table" then
-            if file_attr.mode ~= "directory" then
-                debug("err",conf_tab.path .." is not a directory")
-                return
-            end
-            state = file_attr[conf_tab.check]
-        else
-            state = "not_found"
-            debug("err",conf_tab.path .." not found")
+    local file_attr = lfs.attributes(conf_tab.path)
+    if type(file_attr) == "table" then
+        if file_attr.mode ~= "directory" then
+            debug("err",conf_tab.path .." is not a directory")
+            return
         end
+    else
+        state = "not_found"
+        debug("err",conf_tab.path .." not found")
+    end
+    if conf_tab.check == "modification" then
+        state = state or file_attr[conf_tab.check]
+    elseif conf_tab.check == "size" then
+        if state ~= "not_found" then
+            local du = get_shell_cmd_result("du -s "..conf_tab.path)
+            local s = split(du,"\t",1)
+            state = tonumber(s)
+        end
+    else
+        debug("err","not support check method!exit...")
+        os.exit()
     end
 
     if not G_custom_conf[conf_tab.name]["last_state"] then
@@ -386,7 +463,7 @@ end
 
 --{"name":"demo","type":"process","cmd":"/usr/bin/test","pidfile":"/var/run/test.pid","period":10}
 --{"name":"demo_file","type":"file","check":"modification/md5/size","cmd":"/usr/bin/file_has_change.cb","period":10,"path":"/tmp/haha"}
---{"name":"demo_dir","type":"dir","check":"modification","cmd":"/usr/bin/dir_has_change.cb","period":10,"path":"/tmp/dir_test"}
+--{"name":"demo_dir","type":"dir","check":"modification/size","cmd":"/usr/bin/dir_has_change.cb","period":10,"path":"/tmp/dir_test"}
 --{"name":"demo_timer","type":"timer","cmd":"/usr/bin/timer_exec.sh","period":10}
 function custom_conf_check(conf_tab)
     if not conf_tab.name or not conf_tab.period or not conf_tab.type then
@@ -427,17 +504,20 @@ function custom_conf_add_to_task(conf_tab)
     end
 end
 
+
 function walk_all_conf_json(include_tab)
     if type(include_tab) ~= "table" then
+        debug("err","include_tab err")
         return
     end
     for _,conf_dir in ipairs(include_tab) do
+        if file_exists(conf_dir ) then
         for file in lfs.dir(conf_dir) do
             if file ~= "." and file ~= ".." then
                 local f = conf_dir.."/"..file
                 debug("info","\t parsing "..f.." <=")
                 local conf = readfile(f)
-                local conf_tab = cjson.decode(conf)
+                local conf_tab = cjson_decode_safe(conf)
                 if type(conf_tab) == "table" then
                     custom_conf_check(conf_tab)
                 else
@@ -447,19 +527,26 @@ function walk_all_conf_json(include_tab)
                 custom_conf_add_to_task(conf_tab)
             end
         end
+        end
     end
 end
 
---main
-print("monitor start version:"..MY_VERSION)
---{"include":["dir3","dir1","dir2"],"pidfile":"/var/run/lua-monitor.pid","debug_level":"info/err/warn","debug_file":"/dev/null"}
-DEFAULT_CONF_FILE="/etc/lua-monitor/conf.json"
+function task_cb_logrotate(arg)
+    os.execute("logrotate /etc/logrotate.conf &")
+    debug("info","run logrotate")
+end
 G_conf ={
     include={"/etc/lua-monitor/conf.d"},
     pidfile="/var/run/lua-monitor.pid",
     debug_level="info",
     debug_file=nil
 }
+--main
+print("monitor start version:"..MY_VERSION)
+os.remove(CTRL_MSG_FILE)
+--{"include":["dir3","dir1","dir2"],"pidfile":"/var/run/lua-monitor.pid","debug_level":"info/err/warn","debug_file":"/dev/null"}
+DEFAULT_CONF_FILE="/etc/lua-monitor/conf.json"
+
 function read_default_conf()
     local conf = readfile(DEFAULT_CONF_FILE)
     local conf_tab = cjson_decode_safe(conf)
@@ -473,15 +560,26 @@ function read_default_conf()
     end
 end
 read_default_conf()
+debug("info","config json:"..cjson.encode(G_conf))
 writeData2file(G_conf.pidfile,tostring(get_mypid()))
+--增加自己的logrotate配置文件
+if G_conf.debug_file and "/dev/null" ~= G_conf.debug_file then
+    local myconf={}
+    myconf.name="_monitor_"
+    myconf.log_file = G_conf.debug_file
+    myconf.log_size = "500K"
+    myconf.log_cnt = 5
+    myconf.log_compress = true
+    generate_logrotate_conf(myconf)
+end
 walk_all_conf_json(G_conf.include)
-
+--add logrotate timer
+timer_task_add("timer_logrotate",{func=task_cb_logrotate,period=600,args={}})
 
 while 1 do
     do_timer_task()
     check_ctrl_msg()
-    
-    sleep(1)
+    sleep(0.5)
 end
